Fix and improve `std::endl` article (#22)
![grafik](https://github.com/jeremy-rifkin/wheatley-mirror/assets/22040976/c386c3d1-0681-45ae-8f75-f925bb3885fd)

The current version has a broken SO emote, and could be enriched with a
number of links in the text. This also allows shortening the *See Also*
section.

I'm also making a few grammatical improvements, because I wasn't as good
at English back when I wrote the original, and have made some mistakes.